requirements.txt
requests
python-dotenv
pandas
matplotlib
streamlit
